[cook] remove internal/lo
